feat: make plugins opt-in
