--- test/snapshots/codeblock_rendered_using_default_and_custom_plugins.snap.html
@@ -1,2 +1,0 @@
-<div class="hl"><pre><code tabindex="0" class="language-zsh" style="--hl-line-gutter: 1"><span data-line-number="1">defaults write com.apple.finder AppleShowAllFiles YES</span>
-</code></pre><div class="hl-footer"><div class="hl-language">zsh</div><button class="hl-copy">Copy to clipboard</button></div></div>
