+++ test/snapshots/codeblock_rendered_with_custom_namespace.snap.html
@@ -1,2 +1,2 @@
-<div class="highlight"><pre><code tabindex="0" class="language-java" style="--hl-line-gutter: 1"><span data-line-number="1"><span class="pl-c1">IO</span><span class="pl-k">.</span>println(<span class="pl-s"><span class="pl-pds">"</span>Hello, world!<span class="pl-pds">"</span></span>);</span>
-</code></pre><div class="highlight-footer"><div class="highlight-language">java</div></div></div>
+<div class="highlight"><pre><code tabindex="0" class="language-java"><span class="pl-c1">IO</span><span class="pl-k">.</span>println(<span class="pl-s"><span class="pl-pds">"</span>Hello, world!<span class="pl-pds">"</span></span>);
+</code></pre></div>
