--- test/snapshots/codeblock_rendered_without_plugins.snap.html
@@ -1,2 +1,0 @@
-<div class="hl"><pre><code tabindex="0" class="language-sh">zip function.zip index.mjs
-</code></pre></div>
