+++ test/snapshots/codeblock_with_aliased_language.snap.html
@@ -1,4 +1,4 @@
-<div class="hl"><pre><code tabindex="0" class="language-psql" style="--hl-line-gutter: 1"><span data-line-number="1"><span class="pl-k">select</span> <span class="pl-k">*</span></span>
-<span data-line-number="2"><span class="pl-k">from</span> books</span>
-<span data-line-number="3"><span class="pl-k">where</span> (published_year, available) <span class="pl-k">=</span> (<span class="pl-c1">2024</span>, TRUE);</span>
-</code></pre><div class="hl-footer"><div class="hl-language">psql</div></div></div>
+<div class="hl"><pre><code tabindex="0" class="language-psql"><span class="pl-k">select</span> <span class="pl-k">*</span>
+<span class="pl-k">from</span> books
+<span class="pl-k">where</span> (published_year, available) <span class="pl-k">=</span> (<span class="pl-c1">2024</span>, TRUE);
+</code></pre></div>
