+++ test/snapshots/codeblock_with_command_and_output.snap.html
@@ -4,4 +4,4 @@
 <span data-line-number="4" data-line-output="">Java version: <span class="pl-c1">25</span>.<span class="pl-c1">0</span>.<span class="pl-c1">1</span>, vendor: Azul Systems, Inc., runtime: <span class="pl-k">/</span>Library<span class="pl-k">/</span>Java<span class="pl-k">/</span>JavaVirtualMachines<span class="pl-k">/</span>zulu<span class="pl-k">-</span><span class="pl-c1">25</span>.jdk<span class="pl-k">/</span>Contents<span class="pl-k">/</span>Home</span>
 <span data-line-number="5" data-line-output="">Default locale: en_US, platform encoding: UTF<span class="pl-k">-</span><span class="pl-c1">8</span></span>
 <span data-line-number="6" data-line-output="">OS name: <span class="pl-s"><span class="pl-pds">"</span>mac os x<span class="pl-pds">"</span></span>, version: <span class="pl-s"><span class="pl-pds">"</span>15.7.2<span class="pl-pds">"</span></span>, arch: <span class="pl-s"><span class="pl-pds">"</span>aarch64<span class="pl-pds">"</span></span>, family: <span class="pl-s"><span class="pl-pds">"</span>mac<span class="pl-pds">"</span></span></span>
-</code></pre><div class="hl-footer"><div class="hl-language">sql</div></div></div>
+</code></pre></div>
