+++ test/snapshots/codeblock_with_deleted_lines.snap.html
@@ -8,4 +8,4 @@
 <span data-line-number="8"><span class="diff">	</span>		<span class="pl-smi">Files</span><span class="pl-k">.</span>readAllLines(path)<span class="pl-k">.</span>forEach(<span class="pl-c1">IO</span><span class="pl-k">::</span>println);</span>
 <span data-line-number="9"><span class="diff">	</span>	}</span>
 <span data-line-number="10" data-line-deleted=""><span class="diff">-	</span>}</span>
-</code></pre><div class="hl-footer"><div class="hl-language">java</div></div></div>
+</code></pre></div>
