+++ test/snapshots/codeblock_with_inserted_lines.snap.html
@@ -3,4 +3,4 @@
 <span data-line-number="3"><span class="diff">	</span>	<span class="pl-s"><span class="pl-pds">"</span>/opt/homebrew/bin<span class="pl-pds">"</span></span>,</span>
 <span data-line-number="4" data-line-inserted=""><span class="diff">+	</span>	<span class="pl-smi">$env</span>.PNPM_HOME</span>
 <span data-line-number="5"><span class="diff">	</span>]</span>
-</code></pre><div class="hl-footer"><div class="hl-language">sh</div></div></div>
+</code></pre></div>
