+++ test/snapshots/codeblock_with_known_language.snap.html
@@ -1,4 +1,4 @@
-<div class="hl"><pre><code tabindex="0" class="language-css" style="--hl-line-gutter: 1"><span data-line-number="1"><span class="pl-ent">*</span> {</span>
-<span data-line-number="2">  <span class="pl-c1">all</span>: <span class="pl-c1">unset</span>;</span>
-<span data-line-number="3">}</span>
-</code></pre><div class="hl-footer"><div class="hl-language">css</div></div></div>
+<div class="hl"><pre><code tabindex="0" class="language-css"><span class="pl-ent">*</span> {
+  <span class="pl-c1">all</span>: <span class="pl-c1">unset</span>;
+}
+</code></pre></div>
