+++ test/snapshots/codeblock_with_language_in_all_grammars.snap.html
@@ -1,6 +1,6 @@
-<div class="hl"><pre><code tabindex="0" class="language-ballerina" style="--hl-line-gutter: 1"><span data-line-number="1"><span class="pl-k">import</span> <span class="pl-smi">ballerina</span>/<span class="pl-smi">io</span>;</span>
-<span data-line-number="2"></span>
-<span data-line-number="3"><span class="pl-k">public</span> <span class="pl-k">function</span> <span class="pl-en">main</span>() {</span>
-<span data-line-number="4">  <span class="pl-smi">io</span><span class="pl-k">:</span><span class="pl-en">println</span>(<span class="pl-s"><span class="pl-pds">"</span>Hello, World!<span class="pl-pds">"</span></span>);</span>
-<span data-line-number="5">}</span>
-</code></pre><div class="hl-footer"><div class="hl-language">ballerina</div></div></div>
+<div class="hl"><pre><code tabindex="0" class="language-ballerina"><span class="pl-k">import</span> <span class="pl-smi">ballerina</span>/<span class="pl-smi">io</span>;
+
+<span class="pl-k">public</span> <span class="pl-k">function</span> <span class="pl-en">main</span>() {
+  <span class="pl-smi">io</span><span class="pl-k">:</span><span class="pl-en">println</span>(<span class="pl-s"><span class="pl-pds">"</span>Hello, World!<span class="pl-pds">"</span></span>);
+}
+</code></pre></div>
