+++ test/snapshots/codeblock_with_language_indicator.snap.html
@@ -1,0 +1,4 @@
+<div class="hl"><pre><code tabindex="0" class="language-rust"><span class="pl-k">fn</span> <span class="pl-en">main</span>() {
+    <span class="pl-en">println!</span>(<span class="pl-s"><span class="pl-pds">"</span>Hello, world!<span class="pl-pds">"</span></span>);
+}
+</code></pre><div class="hl-footer"><div class="hl-language">rust</div></div></div>
