+++ test/snapshots/codeblock_with_marked_lines.snap.html
@@ -11,4 +11,4 @@
 <span data-line-number="11">		<span class="pl-s"><span class="pl-pds">"</span>ID<span class="pl-pds">"</span></span>: <span class="pl-s"><span class="pl-pds">"</span>bcaf1ffd86f41161ca5fb16fd081034f<span class="pl-pds">"</span></span></span>
 <span data-line-number="12">	}</span>
 <span data-line-number="13">}</span>
-</code></pre><div class="hl-footer"><div class="hl-language">sh</div></div></div>
+</code></pre></div>
