+++ test/snapshots/codeblock_with_prompts.snap.html
@@ -2,4 +2,4 @@
 <span data-line-number="2">{<span class="pl-s"><span class="pl-pds">"</span>status<span class="pl-pds">"</span></span>:<span class="pl-s"><span class="pl-pds">"</span>UP<span class="pl-pds">"</span></span>}</span>
 <span data-line-number="3"><span class="prompt" aria-hidden="true"></span>curl localhost:8080/greeter<span class="pl-k">?</span>name=Anya</span>
 <span data-line-number="4">Hello, Anya<span class="pl-k">!</span></span>
-</code></pre><div class="hl-footer"><div class="hl-language">sh</div></div></div>
+</code></pre></div>
