+++ test/snapshots/codeblock_with_title.snap.html
@@ -1,5 +1,5 @@
-<div class="hl"><div class="hl-header"><div class="hl-title">Switching off homebrew telemetry</div></div><pre><code tabindex="0" class="language-zsh" style="--hl-line-gutter: 1"><span data-line-number="1"><span class="pl-c"># turns off homebrew telemetry</span></span>
-<span data-line-number="2"><span class="pl-k">export</span> HOMEBREW_NO_ANALYTICS=1</span>
-<span data-line-number="3"><span class="pl-c"># turns off homebrew auto-update</span></span>
-<span data-line-number="4"><span class="pl-k">export</span> HOMEBREW_NO_AUTO_UPDATE=1</span>
-</code></pre><div class="hl-footer"><div class="hl-language">zsh</div></div></div>
+<div class="hl"><div class="hl-header"><div class="hl-title">Switching off homebrew telemetry</div></div><pre><code tabindex="0" class="language-zsh"><span class="pl-c"># turns off homebrew telemetry</span>
+<span class="pl-k">export</span> HOMEBREW_NO_ANALYTICS=1
+<span class="pl-c"># turns off homebrew auto-update</span>
+<span class="pl-k">export</span> HOMEBREW_NO_AUTO_UPDATE=1
+</code></pre></div>
